# Add a flag to use FusedMoE kernel in compressed quantization