Summary: Allows customizing kernel usage in compressed-tensor quantization.

Test Plan:
CUDA_VISIBLE_DEVICES=6,7 \
  VLLM_DISABLE_COMPILE_CACHE=1 \
  VLLM_MQ_MAX_CHUNK_BYTES_MB=256 \
  VLLM_GPU_MEMORY_UTILIZATION=0.85 \
  VLLM_USE_FUSED_MOE_KERNEL_IN_COMPRESSED_QUANTIZATION=1 \
  buck2 run @//mode/{opt,inplace} \
    -c fbcode.enable_vllm=true \
    -c fbcode.enable_gpu_sections=true \
    -c fbcode.nvcc_arch=h100a \
    //smart/inference_platform_sp/llm_predictor_gpu:service -- \
    --local_cache_dir "$HOME/local/models/GLM-4.5V-FP8" \
    --try_local_cache \
    --max_seq_len=16384 \
    --max_batch_size 192 \
    --thrift_server_port 12345 \
    --enable_warmup=true \
    --model_mf_bucket=llm_inference \
    --model_mf_path=tree/oss/GLM-4.5V-FP8 \
    --force_llm_format=true \
    --allow_custom_stop_tokens \
    --model_parallel_size 2 \
    --vllm_engine \
    --cpu_offload_gb=0 \
    --kv_cache_quantization 8

Before:

QPS:                 1.26
Avg latency:         49.998s
Avg TTFT (client):   1679.44ms
P50 TTFT (client):   1584.17ms
P99 TTFT (client):   5748.46ms
Avg TTIT (client):   48.32ms
P50 TTIT (client):   48.21ms
P99 TTIT (client):   59.81ms
Avg TTFT (server):   2481.96ms
Avg TTIT (server):   48.06ms
Avg prefill len:     2643.00 tokens
P50 prefill len:     2643.00 tokens
P99 prefill len:     2643.00 tokens
Avg decode len:      1000.00 tokens
P50 decode len:      1000.00 tokens
P99 decode len:      1000.00 tokens

After:

QPS:                 1.86
Avg latency:         92.238s
Avg TTFT (client):   1856.53ms
P50 TTFT (client):   1912.39ms
P99 TTFT (client):   2694.97ms
Avg TTIT (client):   70.38ms
P50 TTIT (client):   75.11ms
P99 TTIT (client):   76.74ms
Avg TTFT (server):   2984.87ms
Avg TTIT (server):   77.98ms
Avg prefill len:     2643.00 tokens
P50 prefill len:     2643.00 tokens
P99 prefill len:     2643.00 tokens
Avg decode len:      1000.00 tokens
P50 decode len:      1000.00 tokens
P99 decode len:      1000.00 tokens

Rollback Plan:

Differential Revision: D80552023